pyproject.toml
[project]
name = "kids-story-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.19.0",
    "google-genai>=1.52.0",
    "openai>=2.8.1",
    "pillow>=12.0.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "pydantic>=2.10.4",
]
